+++ CustomCraftingStations/docs/release-notes.md
@@ -2,6 +2,7 @@
 
 # Release notes
 ## Upcoming release
+- Added in-game config UI through [Generic Mod Config Menu](https://www.nexusmods.com/stardewvalley/mods/5098).
 - Added validation error if a custom station has both cooking and crafting recipes, which isn't supported by the vanilla menu used since 1.2.0.
 - Fixed cooking recipes in recent versions.