+++ CustomizeAnywhere/docs/release-notes.md
@@ -2,7 +2,8 @@
 
 # Release notes
 ## Upcoming release
-* Improved translations. Thanks to BlackRosePetals (added Chinese)!
+- Added in-game config UI through [Generic Mod Config Menu](https://www.nexusmods.com/stardewvalley/mods/5098).
+- Improved translations. Thanks to BlackRosePetals (added Chinese)!
 
 ## 1.3.0
 Released 21 January 2025 for SMAPI 4.1.10 or later. Updated by Pathoschild.